refactor: extract members app from blowcomotion monolith
