+++ blowcomotion/settings/base.py
@@ -34,6 +34,7 @@ INSTALLED_APPS = [
     "attendance",
     "charts",
     "instruments",
+    "members",
     "search",
     "compressor",
     "livereload",
@@ -69,7 +70,7 @@ MIDDLEWARE = [
     "django.middleware.csrf.CsrfViewMiddleware",
     "django.contrib.auth.middleware.AuthenticationMiddleware",
     "axes.middleware.AxesMiddleware",
-    "blowcomotion.middleware.MemberIdleLogoutMiddleware",
+    "members.middleware.MemberIdleLogoutMiddleware",
     "django.contrib.messages.middleware.MessageMiddleware",
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
     "django.middleware.security.SecurityMiddleware",
