+++ members/templates/emails/member_signup_invite.txt
