+++ members/templates/member/password_reset_email.txt
